requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
defusedxml==0.7.1
Django==5.0.7
django-cors-headers==4.4.0
django-unused-media==0.2.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
httpie==3.2.3
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
mysqlclient==2.2.4
numpy==2.1.0
phonenumbers==8.13.44
pillow==10.4.0
psycopg2==2.9.9
Pygments==2.18.0
PyJWT==2.8.0
PySocks==1.7.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
